+++ lib/src/endpoints/users.dart
@@ -1,14 +1,13 @@
 // Copyright (c) 2017, chances. All rights reserved. Use of this source code
 // is governed by a BSD-style license that can be found in the LICENSE file.
 
-part of spotify;
+part of '../../spotify.dart';
 
 class Users extends EndpointPaging {
-
   @override
   String get _path => 'v1/users';
 
-  Users(SpotifyApiBase api) : super(api);
+  Users(super.api);
 
   Future<UserPublic> get(String userId) async {
     var jsonString = await _api._get('$_path/$userId');