+++ lib/src/models/copyright.dart
@@ -1,7 +1,7 @@
 // Copyright (c) 2017, rinukkusu. All rights reserved. Use of this source code
 // is governed by a BSD-style license that can be found in the LICENSE file.
 
-part of spotify.models;
+part of '_models.dart';
 
 /// Json representation of copyright
 @JsonSerializable(createToJson: false)
@@ -21,6 +21,7 @@ class Copyright extends Object {
 enum CopyrightType {
   /// C = the copyright
   C,
+
   /// P = the sound recording (performance) copyright.
   P
 }