+++ lib/src/models/error.dart
@@ -1,7 +1,7 @@
 // Copyright (c) 2018, chances. All rights reserved. Use of this source code
 // is governed by a BSD-style license that can be found in the LICENSE file.
 
-part of spotify.models;
+part of '_models.dart';
 
 /// Json representation of an API error
 @JsonSerializable(createToJson: false)
@@ -13,6 +13,7 @@ class SpotifyError extends Object {
 
   /// The HTTP status code (also returned in the response header; see Response
   /// Status Codes for more information).
+  @JsonKey(fromJson: convertToIntIfDoubleValue)
   int? status;
 
   /// A short description of the cause of the error.