+++ lib/src/spotify_base.dart
@@ -1,7 +1,7 @@
 // Copyright (c) 2017, 'rinukkusu'. All rights reserved. Use of this source code
 // is governed by a BSD-style license that can be found in the LICENSE file.
 
-part of spotify;
+part of '../spotify.dart';
 
 abstract class SpotifyApiBase {
   static const String _baseUrl = 'https://api.spotify.com';
@@ -10,40 +10,57 @@ abstract class SpotifyApiBase {
       'https://accounts.spotify.com/authorize';
 
   bool _shouldWait = false;
+
   late FutureOr<oauth2.Client> _client;
+  FutureOr<oauth2.Client> get client => _client;
+
   late Artists _artists;
   Artists get artists => _artists;
+
   late Albums _albums;
   Albums get albums => _albums;
+
   late Browse _browse;
   Browse get browse => _browse;
+
   late Tracks _tracks;
   Tracks get tracks => _tracks;
+
   late Playlists _playlists;
   Playlists get playlists => _playlists;
+
   late Episodes _episodes;
   Episodes get episodes => _episodes;
+
   late RecommendationsEndpoint _recommendations;
   RecommendationsEndpoint get recommendations => _recommendations;
+
   late Markets _markets;
   Markets get markets => _markets;
+
   late Users _users;
   Users get users => _users;
+
   late Search _search;
   Search get search => _search;
+
   late AudioFeatures _audioFeatures;
   AudioFeatures get audioFeatures => _audioFeatures;
+
   late AudioAnalysisEndpoint _audioAnalysis;
   AudioAnalysisEndpoint get audioAnalysis => _audioAnalysis;
+
   late Categories _categories;
   Categories get categories => _categories;
+
   late Me _me;
   Me get me => _me;
+
   late PlayerEndpoint _player;
   PlayerEndpoint get player => _player;
+
   late Shows _shows;
   Shows get shows => _shows;
-  FutureOr<oauth2.Client> get client => _client;
   interceptor.Client _interceptor;
 
   SpotifyApiBase.fromClient(FutureOr<oauth2.Client> client) {
@@ -98,14 +115,16 @@ abstract class SpotifyApiBase {
 
   static oauth2.AuthorizationCodeGrant authorizationCodeGrant(
       SpotifyApiCredentials credentials, http.Client httpClient,
-      [Function(SpotifyApiCredentials)? callBack]) {
+      {String? codeVerifier,
+      Function(SpotifyApiCredentials)? onCredentialsRefreshed}) {
     return oauth2.AuthorizationCodeGrant(
         credentials.clientId!,
         Uri.parse(SpotifyApiBase._authorizationUrl),
         Uri.parse(SpotifyApiBase._tokenUrl),
         secret: credentials.clientSecret,
+        codeVerifier: codeVerifier,
         httpClient: httpClient,
-        onCredentialsRefreshed: callBack != null
+        onCredentialsRefreshed: onCredentialsRefreshed != null
             ? (oauth2.Credentials cred) {
                 final newCredentials = SpotifyApiCredentials(
                     credentials.clientId, credentials.clientSecret,
@@ -113,7 +132,7 @@ abstract class SpotifyApiBase {
                     expiration: cred.expiration,
                     refreshToken: cred.refreshToken,
                     scopes: cred.scopes);
-                callBack(newCredentials);
+                onCredentialsRefreshed(newCredentials);
               }
             : null);
   }
@@ -178,7 +197,7 @@ abstract class SpotifyApiBase {
     return _deleteImpl('$_baseUrl/$path', const {}, body);
   }
 
-  Future<String> _put(String path, [String body = '']) {
+  Future<String> _put(String path, [String? body]) {
     return _putImpl('$_baseUrl/$path', const {}, body);
   }
 
@@ -230,13 +249,13 @@ abstract class SpotifyApiBase {
       }
       try {
         var response = await request();
-
+        
         // distinguish between url redirect responses and body responses
         // note, that any response that also contains a redirect url
         // will be chosen instead of its body contents
         // FIXME: in future releases of http2, the url is a part of the [http.Response] type
         if (response case http.BaseResponseWithUrl(:final url)) {
-          return url.toString();
+            return url.toString();
         }
         return handleResponseWithBody(response as http.Response);
       } on ApiRateException catch (ex) {