+++ lib/src/spotify_exception.dart
@@ -1,7 +1,7 @@
 // Copyright (c) 2018, chances. All rights reserved. Use of this source code
 // is governed by a BSD-style license that can be found in the LICENSE file.
 
-part of spotify;
+part of '../spotify.dart';
 
 class SpotifyException implements Exception {
   int? status;
@@ -22,6 +22,6 @@ class SpotifyException implements Exception {
 class ApiRateException extends SpotifyException {
   final num retryAfter;
 
-  ApiRateException.fromSpotify(SpotifyError error, this.retryAfter)
-      : super.fromSpotify(error);
+  ApiRateException.fromSpotify(super.error, this.retryAfter)
+      : super.fromSpotify();
 }